Ok this version works.
What it does is it locks your HDD upon load of the Chimp menu and then unlocks it, if it fails to lock the drive it will notify you that the drive cannot be locked.
So this is a sure way to tell if your HDD is lockable, as I try to lock it ;) and if it fails well its not lockable.

If for what ever reason you get this notification and the drive is lockable let me know, so I can fix it and send you a test build to test for me.
( I have no drives that aren't lockable )