Another small update - 17/12/2016:
Well it would seem I have to change the menu system because people still unlock there softmodded Xbox HDD when the options are disabled by default.
So I have added a new menu and a disclaimer that you must agree to when you pick an option.
I also disabled options based on the mod type you select and also made the lock HDD error a but more pleasing to the eye.

Note: If you get the slave drive error about locking, this is because HDTool cannot lock the drive, there is nothing I can do to fix this.