Redone the Partition information screen, now shows information in a nice manor.
Added a check to see if its a CD-ROM or a hard drive when loading.
( this stops any errors appearing if you are using chimp with a DVD drive connected )

Added a check for maximum lock status when selecting clone from the menu.

Two new scripts for displaying the partition information for the drives.

also other odds and ends fixed/amended.